pyproject.toml
[project]
name = "protocols"
version = "0.1.0"
description = "Understanding MCP & A2A protocols"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a>=0.44",
    "agno>=2.3.21",
    "aiosqlite>=0.22.1",
    "arize-phoenix-otel>=0.14.0",
    "fastapi[standard]>=0.115.12",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "langgraph>=0.4.7",
    "mcp>=1.9.1",
    "openai>=2.14.0",
    "openinference-instrumentation-langchain>=0.1.57",
    "ruff>=0.11.11",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.2",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "langgraph-cli[inmem]>=0.4.11",
    "pre-commit>=4.5.1",
    "yamlfix>=1.19.1",
]
